pyproject.toml
[tool.poetry]
name = "mattermost-meetbot"
version = "0.1.0"
description = ""
authors = ["Bartlomiej Biernacki <pax0r@o2.pl>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
google-api-python-client = "^2.48.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.1"
uvicorn = "^0.17.6"
redis = "^4.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
